Add files via upload
This script seeks to uninstall large portions of "Windows 11 bloat" that comes pre-installed. I don't really use any of these applications so I choose to remove them. The script will confirm to the user if they want to uninstall the application or not (y/n). This is just a list I threw together and may not be all encompassing. 

Also, a reminder that you cannot uninstall the Windows Start menu shortcuts to applications such as: "tiktok", "ESPN", or "Instagram". You still need to "uninstall" those by right clicking and clicking "uninstall" manually. These are not really installed. They will merely install the application if you click on it the first time.